+++ listenbrainz/listenstore/timescale_listenstore.py
@@ -242,7 +242,7 @@ class TimescaleListenStore:
                         , mm.recording_mbid
                         , mbc.release_mbid
                         , mbc.artist_mbids
-                        , (mbc.release_data->>'caa_id')::int
+                        , (mbc.release_data->>'caa_id')::bigint
                         , mbc.release_data->>'caa_release_mbid'
                      FROM listen
                 LEFT JOIN mbid_mapping mm
@@ -372,7 +372,7 @@ class TimescaleListenStore:
                                     , mm.recording_mbid
                                     , mbc.release_mbid
                                     , mbc.artist_mbids
-                                    , (mbc.release_data->>'caa_id')::int
+                                    , (mbc.release_data->>'caa_id')::bigint
                                     , mbc.release_data->>'caa_release_mbid'
                                     , row_number() OVER (PARTITION BY user_id ORDER BY listened_at DESC) AS rownum
                                 FROM listen l
